[CORE] Js: converter system
